WIP commit. Moved assertion class to separate package.
