+++ src/test/java/org/junitpioneer/jupiter/PioneerAnnotationUtilsTests.java
@@ -10,7 +10,7 @@
 
 package org.junitpioneer.jupiter;
 
-import static org.junitpioneer.testkit.PioneerAssertContainer.PioneerAssert.assertThat;
+import static org.junitpioneer.testkit.assertion.PioneerAssert.assertThat;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
