+++ src/test/java/org/junitpioneer/testkit/assertion/SingleReportEntryAssert.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2015-2020 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v2.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v20.html
+ */
+
+package org.junitpioneer.testkit.assertion;
+
+/**
+ * Used to assert a single report entry.
+ */
+public interface SingleReportEntryAssert {
+
+	void withKeyAndValue(String key, String value);
+
+}
